Added melee weapon & a car script
Added a bat that is connected the the players hand. an animation and collider so you can hit stuff. and a test cube to test so the bat works. The bat still need to only do damage when you are hitting the object. right now it adds damage also when the colliders softly touch.
Added DanielsCarScript for car movements. also added a prefab for the car named DanielSedan.